Cleanups & renames
Update README.md

Cleaner bucketMap and new tests

Less looping

bucket.allow was unused (by the limiter)

Move some files around

Rename Debug and add benchmarks
