+++ go.mod
@@ -7,6 +7,5 @@ require github.com/stretchr/testify v1.10.0
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
-	golang.org/x/time v0.12.0
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
