+++ limiter_bench_test.go
@@ -45,12 +45,12 @@ func BenchmarkLimiter_Allow_MultipleBuckets(b *testing.B) {
 	b.Run("SingleLimit", func(b *testing.B) {
 		limit := NewLimit(1000000, time.Second)
 		limiter := NewLimiter(keyer, limit)
-		now := time.Now()
 
 		b.ResetTimer()
 		b.ReportAllocs()
 
-		for i := 0; i < b.N; i++ {
+		for i := 0; b.Loop(); i++ {
+			now := time.Now()
 			limiter.allowN(i%buckets, now, 1)
 		}
 	})
@@ -64,7 +64,7 @@ func BenchmarkLimiter_Allow_MultipleBuckets(b *testing.B) {
 		b.ResetTimer()
 		b.ReportAllocs()
 
-		for i := 0; i < b.N; i++ {
+		for i := 0; b.Loop(); i++ {
 			limiter.allowN(i%buckets, now, 1)
 		}
 	})
@@ -245,7 +245,7 @@ func BenchmarkLimiter_PeekWithDetails_MultipleBuckets(b *testing.B) {
 		b.ResetTimer()
 		b.ReportAllocs()
 
-		for i := 0; i < b.N; i++ {
+		for i := 0; b.Loop(); i++ {
 			limiter.peekNWithDetails(i%buckets, now, 1)
 		}
 	})
@@ -259,8 +259,138 @@ func BenchmarkLimiter_PeekWithDetails_MultipleBuckets(b *testing.B) {
 		b.ResetTimer()
 		b.ReportAllocs()
 
-		for i := 0; i < b.N; i++ {
+		for i := 0; b.Loop(); i++ {
 			limiter.peekNWithDetails(i%buckets, now, 1)
+		}
+	})
+}
+
+func BenchmarkLimiter_AllowWithDebug_SingleBucket(b *testing.B) {
+	keyer := func(id int) string {
+		return "allow-debug-single-bucket"
+	}
+
+	b.Run("SingleLimit", func(b *testing.B) {
+		limit := NewLimit(1000000, time.Second)
+		limiter := NewLimiter(keyer, limit)
+		now := time.Now()
+
+		b.ReportAllocs()
+
+		for b.Loop() {
+			limiter.allowNWithDebug(0, now, 1)
+		}
+	})
+
+	b.Run("MultipleLimits", func(b *testing.B) {
+		limit1 := NewLimit(1000000, time.Second)
+		limit2 := NewLimit(500000, time.Second/2)
+		limiter := NewLimiter(keyer, limit1, limit2)
+		now := time.Now()
+
+		b.ReportAllocs()
+
+		for b.Loop() {
+			limiter.allowNWithDebug(0, now, 1)
+		}
+	})
+}
+
+func BenchmarkLimiter_AllowWithDebug_MultipleBuckets(b *testing.B) {
+	const buckets = 1000
+	keyer := func(id int) int {
+		return id
+	}
+
+	b.Run("SingleLimit", func(b *testing.B) {
+		limit := NewLimit(1000000, time.Second)
+		limiter := NewLimiter(keyer, limit)
+		now := time.Now()
+
+		b.ResetTimer()
+		b.ReportAllocs()
+
+		for i := 0; b.Loop(); i++ {
+			limiter.allowNWithDebug(i%buckets, now, 1)
+		}
+	})
+
+	b.Run("MultipleLimits", func(b *testing.B) {
+		limit1 := NewLimit(1000000, time.Second)
+		limit2 := NewLimit(500000, time.Second/2)
+		limiter := NewLimiter(keyer, limit1, limit2)
+		now := time.Now()
+
+		b.ResetTimer()
+		b.ReportAllocs()
+
+		for i := 0; b.Loop(); i++ {
+			limiter.allowNWithDebug(i%buckets, now, 1)
+		}
+	})
+}
+
+func BenchmarkLimiter_PeekWithDebug_SingleBucket(b *testing.B) {
+	keyer := func(id int) string {
+		return "peek-debug-single-bucket"
+	}
+
+	b.Run("SingleLimit", func(b *testing.B) {
+		limit := NewLimit(1000000, time.Second)
+		limiter := NewLimiter(keyer, limit)
+		now := time.Now()
+
+		b.ReportAllocs()
+
+		for b.Loop() {
+			limiter.peekNWithDebug(0, now, 1)
+		}
+	})
+
+	b.Run("MultipleLimits", func(b *testing.B) {
+		limit1 := NewLimit(1000000, time.Second)
+		limit2 := NewLimit(500000, time.Second/2)
+		limiter := NewLimiter(keyer, limit1, limit2)
+		now := time.Now()
+
+		b.ReportAllocs()
+
+		for b.Loop() {
+			limiter.peekNWithDebug(0, now, 1)
+		}
+	})
+}
+
+func BenchmarkLimiter_PeekWithDebug_MultipleBuckets(b *testing.B) {
+	const buckets = 1000
+	keyer := func(id int) int {
+		return id
+	}
+
+	b.Run("SingleLimit", func(b *testing.B) {
+		limit := NewLimit(1000000, time.Second)
+		limiter := NewLimiter(keyer, limit)
+		now := time.Now()
+
+		b.ResetTimer()
+		b.ReportAllocs()
+
+		for i := 0; b.Loop(); i++ {
+			limiter.peekNWithDebug(i%buckets, now, 1)
+		}
+	})
+
+	b.Run("MultipleLimits", func(b *testing.B) {
+		limit1 := NewLimit(1000000, time.Second)
+		limit2 := NewLimit(500000, time.Second/2)
+		limiter := NewLimiter(keyer, limit1, limit2)
+		now := time.Now()
+
+		b.ResetTimer()
+		b.ReportAllocs()
+
+		for i := 0; b.Loop(); i++ {
+			limiter.peekNWithDebug(i%buckets, now, 1)
 		}
 	})
 }
